[JTF] The Final Refactor
Part one of the final refator. We down to 1100 lines of code in the main class (Woot Woot)
Finished all of the ciphers except for the last one 'Substitute Solver' --it is the biggest fish to fry it seems like.
adding 9 Cipher classes

created a CookieCutterViginereCipher -- this class extends BaseCipher and it has subclasses that has Ciphers that look similiar to the look of the Viginere Cipher. Just a way to reduce code resuse in 6 different files since it was not necessary. Not the most useful thing to do but I wanted to have extends inception so to speak...
added 3 new methods to CipherService. One of which is being used (calculateFriedman) the other two new ones are currently not being used. Might be able to use one of them. Just testing.