[JTF] The Final Refactor #27
Finished out all of the refactoring for the ciphers class. That part is complete. Some more refactoring can be done with respect to the main file but I believe that this will be more than sufficient for now; we can make the main file, as well as the constructors in the Ciphers more clean and refactor those if we get bored/run out of things to do.
Main file is down to ~800 lines which is hype.

Created an OptionsMenu class which just extracts out the menu into its own class