Starting the Hill Cipher
I made a new class called Key that I will use when creating the encryption and decryption for this cipher. I will work on this at a slower pace because of the ciphers complexity. I also made a Junit test for the class to test the creation and the key creation.